+++ examples/Protobuf/Program.cs
@@ -32,9 +32,9 @@ namespace Confluent.Kafka.Examples.Protobuf
     /// <summary>
     ///     protobuf serializer
     /// </summary>
-    public class ProtobufSerializer<T> : ISerializer<T> where T : IMessage<T>, new()
+    public class ProtobufSerializer<T> : ISimpleSerializer<T> where T : IMessage<T>, new()
     {
-        public byte[] Serialize(T data, SerializationContext context)
+        public byte[] Serialize(T data)
             => data.ToByteArray();
     }
 
@@ -50,7 +50,7 @@ namespace Confluent.Kafka.Examples.Protobuf
             parser = new MessageParser<T>(() => new T());
         }
 
-        public T Deserialize(ReadOnlySpan<byte> data, bool isNull, SerializationContext context)
+        public T Deserialize(ReadOnlySpan<byte> data, bool isNull)
             => parser.ParseFrom(data.ToArray());
     }